Moved version code block directive
